reafctor: Refactor payment handling and UI updates in demo app
- Updated ExitHeadlessCallBackManager to use PaymentSheet.PaymentResult.
- Enhanced HyperFragment to handle confirmCVCAction and payment results more effectively.
- Modified HyperHeadlessModule and HyperModule to accommodate new payment result handling.
- Refactored HyperswitchBoundElement to support new payment confirmation methods.
- Improved HyperswitchElement to manage payment results and configurations.
- Updated PaymentWidgetView to streamline payment result handling and configuration.
- Added new drawable for rounded button outline and updated widget_activity layout for better UI.
- Enhanced WidgetActivity to utilize new Hyperswitch SDK features and improve button interactions.

+++ app/src/main/kotlin/io/hyperswitch/react/HyperModule.kt
@@ -161,7 +161,7 @@ class HyperModule internal constructor(private val rct: ReactApplicationContext)
 
     // Method to exit widget payment sheet
     @ReactMethod
-    fun exitWidgetPaymentsheet(rootTag: Int, paymentResult: String, reset: Boolean) {
+    fun exitWidgetPaymentsheet(rootTag: Double, paymentResult: String, reset: Boolean) {
         findFragmentWithRootTag(rootTag.toInt(), {
             it?.notifyResult(CallbackType.PAYMENT_RESULT, paymentResult)
         })
@@ -198,7 +198,6 @@ class HyperModule internal constructor(private val rct: ReactApplicationContext)
 
         }
     }
-
     // Variable to keep track of event listener count
     private val listenerCount = AtomicInteger(0)
